move ActivitySources to their own files
